[sc-15461] Fix compile script
Previous script had two issues. First it attempted to run the task
mode:... which does not exist, real name is ml_model.
Second it attempted to invoke bin/rails which requires the current
directory to be /app which it is not always. Instead we should invoke
rake and run the task that way instead of directly through bin/rails.