[sc-15461] Fix compile script
Rake is still not finding the tasks file so move to the correct
directory to help it locate the Rakefile we are referencing.
We also need to populate the environment variables since we are
attempting to access S3 and need to authenticate using values saved in
the environment vars.
Set e and pipefail to ensure we fail fast and dont lose any error codes
to pipes.
We lastly need to create the directory for the model to get saved to
since the task will fail if the directory doesn't exist.