[sc-15461] Revert compile script
There is a native feature heroku provides called Release Phase which is
designed to do exactly what we are trying to do here. So moving the
logic to download a custom model there and cleaning this back to the
original working version. Also remove a dead variable and switch the
shebang to be the preferred version.