[sc-15461] Rework detect logic
Original checked the length of the Gemfile and assumed a gemfile with
some non-zero length meant rtesseract was installed.
This is silly.
Now just grep the Gemfile for instances of rtesseract and count the
number of resulting lines.
If there are any then rtesseract is probably included and we should
install the buildpack.