[sc-15461] Remove unneeded logic
We dont need to check if a Gemfile exists since catting a file which
does not exist will never yield a string which contains 'rtesseract' and
even if it does, cat erroring will print the error message to stderr so
there will be nothing piped to grep if the file does not exist. We can
thus be certain that if the file does not exist grep will receive
nothing, return nothing, and wc -l will return 0.