[sc-15461] Check if task exists before running
Previously would run task even if it didnt exist which could cause a
failed build if the task was not present.
Now we check if the task exists and only try to run it if it does.
This may cause issues if assumptions are made about the existance of a
custom model but will cause fewer build failures.